+++ src/services/test/dynamo-database-service.test.ts
@@ -19,6 +19,7 @@ jest.mock('@smithy/node-http-handler');
 const ddbMock = mockClient(DynamoDBClient);
 const queryCommandMock = ddbMock.on(QueryCommand);
 const updateCommandMock = ddbMock.on(UpdateItemCommand);
+
 describe('Dynamo DB Service', () => {
   beforeAll(() => {
     jest.useFakeTimers();
@@ -28,7 +29,6 @@ describe('Dynamo DB Service', () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
-
 
   afterEach(() => {
     jest.clearAllMocks();
@@ -153,6 +153,16 @@ describe('Dynamo DB Service', () => {
     const dynamoDBService = new DynamoDatabaseService('table_name')
     await dynamoDBService.updateDeleteStatus('hello');
     expect(ddbMock).toHaveReceivedCommandWith(UpdateItemCommand, commandInput);
+  });
 
+  it('should throw an error if response of the Dynamo Client is null / undefined', async () => {
+    const mockedUpdateCommand = mockClient(DynamoDBClient).on(UpdateItemCommand);
+    mockedUpdateCommand.resolves(undefined as any);
+    const loggerErrorSpy = jest.spyOn(logger, 'error');
+    const dynamoDBService = new DynamoDatabaseService('table_name')
+    await dynamoDBService.updateDeleteStatus('hello');
+    expect(loggerErrorSpy).toHaveBeenCalledWith('DynamoDB may have failed to update items, returned a null response.');
+    expect(logAndPublishMetric).toHaveBeenCalledWith('DB_UPDATE_ERROR');
   });
 });
+
